add: slider folder
