+++ app/page.tsx
@@ -1,5 +1,5 @@
 "use client";
-import TopSlider from "./_components/home/TopSlider";
+import TopSlider from "./_components/home/slider/TopSlider";
 import { useClient } from "./_hooks/useClient";
 import { products } from "./_mock";
